requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlmodel>=0.0.14
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
